create db folder and move prisma files inside
