+++ package.json
@@ -59,7 +59,8 @@
     "vite": "^5.1.4"
   },
   "prisma": {
-    "seed": "tsx prisma/seed.ts"
+    "schema": "./db/prisma/schema.prisma",
+    "seed": "tsx ./db/prisma/seed.ts"
   },
   "engines": {
     "node": ">=18.0.0"
